requirements.txt
# Core Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0
xgboost>=1.5.0
catboost>=1.0.0
joblib>=1.1.0

# MLflow Tracking
mlflow>=2.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Big Data Processing
pyspark>=3.2.0

# Kafka Streaming
confluent-kafka>=1.8.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
statsmodels>=0.13.0
ipython>=7.0.0

# Jupyter Notebook
jupyter>=1.0.0
notebook>=6.4.0